перенес начальную загрузку данных для компонентов в файл App.jsx, добавил в Catalog форму поиска (пока статично), обновил доку
