+++ README.md
@@ -2,14 +2,16 @@
 
 Окружение
 
-Файл api.rest в папке бэка навел на мысль, что нужно установить расширение REST Client для VS Code, с этим плагином можно запускать запросы к бэку прямо в окне редактора с открытым файлом api.rest   
+Файл api.rest в папке бэка навел на мысль, что нужно установить расширение REST Client для VS Code, с этим плагином можно запускать запросы к бэку прямо в окне редактора с открытым файлом api.rest
 
-Требование про правила линтера airbnb отстало от современности - вышел ESLint 9  с новым форматом конфига и всё сломалось.  
+Требование про правила линтера airbnb отстало от современности - вышел ESLint 9 с новым форматом конфига и всё сломалось.  
 Тем не менее, в проекте есть eslint с плагинами для реакта и prettier с рекомендованными и базовыми правилами соответственно.  
 В VS Code включена настройка автоформатирования при выходе - "editor.formatOnSave": true  
 Форматирование меняет код по правилам, eslint при запуске ругается на код, если ошибки есть, VS Code ошибки линтера подсвечивает.
 
 Топик с SoF про airbnb: https://stackoverflow.com/questions/74925642/how-to-use-eslint-config-airbnb-with-the-new-eslint-config-spec
+
+Команды для установки окружения:
 
 ```sh
 cd frontend
@@ -25,7 +27,14 @@ npm i
 npm audit fix
 ```
 
+Команды для установки необходимых пакетов:
+
+```sh
+npm i redux react-redux redux-saga axios
+```
+
 ESLint & Prettier old stuff:
+
 - https://dev.to/suprabhasupi/learn-to-configure-eslint-and-prettier-in-react-4gp0
 - https://github.com/prettier/eslint-plugin-prettier
 - https://prettier.io/docs/precommit
@@ -49,34 +58,99 @@ ESLint & Prettier old stuff:
 - https://mwh-deploy.netlify.app/blog/react-eslint-airbnb-prettier
 - https://stackoverflow.com/questions/71020035/setup-vite-template-react-and-eslint-airbnb
 
-
 ### Иконки для сайта
+
 Создана favicon, сайт https://realfavicongenerator.net/
 
 ### Bootstrap
+
 В версиях Bootstrap5 и Bootstrap 4.3.1 (из проекта) есть различия, пришлось использовать ту версию, которая указана в проекте.  
 Хотя надо обновляться хотя бы до 4.6.2 или мигрировать на v5.
 
-
 ### Routing
-С библиотекой react-router-dom@6 реализован роутинг - компоненты шапки сайта, баннера и подвала не перезагружаются, компоненты разных маршрутов отрисовываются внутри тега \<main\> с помощью элемента Outlet из баблиотеки роутинга.
 
+С библиотекой react-router-dom@6 реализован роутинг - компоненты шапки сайта, баннера и подвала не перезагружаются, компоненты разных маршрутов отрисовываются внутри тега \<main\> с помощью элемента Outlet из баблиотеки роутинга.
 
 ### Создание статических компонентов
 
 #### Header
+
 Реализовано появление/скрытие поля поиска при клике по иконке поиска.  
 Реализован переход на страницу корзины при клике на иконку корзины.  
 При переходе по любым маршрутам (при клике на NavLink), поле поиска скрывается и восстанавливается изначальное состояние.
 
-
 #### Footer
-В представленной верстке есть ошибка, ширина футера визуально больше, чем у хидера, исправил.  
-Открытый адрес электронной почты - это подарок ботам и спамерам. 
 
+В представленной верстке есть ошибка, ширина футера визуально больше, чем у хидера, исправил.  
+Открытый адрес электронной почты - это подарок ботам и спамерам.
 
 #### 404
+
 Реализован компонент, отображающийся при переходе по неверному маршруту (404)
 
+### Работа с состоянием
+
+Будем использовать Redux, Redux Toolkit и Redux-Saga, со стором, сагами и редьюсерами.
+
+### Адрес бэкенда и эндпойнтов API
+Адрес бэкенда описан в .env-файле.  
+Адреса эндпойнтов описаны в файле @/backend_urls.js в проекте, в этом же файле используется адрес бэкенда из переменной окружения, которая формируется из .env-файла.
+
+
+### Обработка ошибок от сервера
+В сагах использован оператор retry для повтора запроса к api в случае ошибки.  
+Количество повторов и время между повторами запросов описано в .env-файле.  
+
+В функции-обработчике запроса к api использован axios и его параметр конфига, отвечающий за максимальное время ожидание ответа от сервера.  
+Пороговое время ожидания ответа от сервера также описано в .env-файле.
 
 #### Хиты продаж
+
+Ответ сервера:
+
+```js
+[
+  {
+    "id": 65,
+    "category": 15,
+    "title": "Босоножки 'Keira'",
+    "price": 7600,
+    "images": [
+      "https://raw.githubusercontent.com/netology-code/ra16-diploma/master/html/img/products/sandals_keira.jpg",
+      "https://raw.githubusercontent.com/netology-code/ra16-diploma/master/html/img/products/sandals_keira_2.jpg"
+     ]
+  },
+  ...
+]
+```
+
+Из верстки выделен компонент карточки товара, обнаружены недостатки верстки, добавлены дополнительные классы.
+
+Немного странно, но пришлось выполнять задание "необходимо распарсить и вывести элементы. Если в ответе пришёл пустой массив, то есть хитов продаж нет, то компонент не должен ничего отображать, как и не должен занимать места на экране", это породило дублирование верстки.
+
+**ВНИМАНИЕ - поправить стили, чтобы картинка товара не растягивалась на весь контейнер!**
+
+
+### Каталог
+Каталог состоит из двух элементов, требующих своих сторов - меню категорий и список товаров
+
+#### Категории
+бэк отдает объекты категорий в массиве
+[
+  {
+    "id": 12,
+    "title": "Мужская обувь"
+  },
+]
+В начальном состоянии есть объект {id: 99999, title: "Все"}, к нему добавляем ответ сервера.  
+Клик по категории должен вести на страницу каталога с фильтрацией с выводом товаров из нужной категории.  Объект категории с id=999 - это хардкод, который может где-то в коде аукнуться, но считаем, что у нас в магазине заведомо не будет 99999 категорий товаров.
+
+#### Каталог товаров
+
+Основная работа в редьюсере, в сторе список не перезаписывается, а конкатенируется.  
+Реализована переменная для признака того, отображать ли кнопку "Загрузить еще".  
+По условию задачи бэк отдает по 6 объектов товаров в списке, но это количество я вынес в переменную в .env-файл.  
+Реализована обработка кейса, когда при загрузке данных для каталога происходит сбой, который не устраняется автоматическими повторами запроса в саге, в таком случае вместо кнопки "Загрузить ещё"  отобразится кнопка "Повторить запрос" с аналогичным функционалом. Это проверено через консоль разработчика в браузере.
+
+### Централизованная загрузка данных
+По ходу реализации пришлось сделать централизованную загрузку данных для TopSales и Catalog в компоненте App.jsx, чтобы при переходе по разным маршрутам приложения не происходило лишних запросов к бэку, что может привести к дублированию данных.
