+++ frontend/.env.sample
@@ -1,0 +1,5 @@
+VITE_BACKEND_URL=http://localhost:7070
+VITE_REQUEST_MAX_ATTEMPTS=3
+VITE_REQUEST_TIMEOUT_MS=10000
+VITE_RETRY_DELAY_MS=2000
+VITE_PRODUCTS_PER_REQUEST=6
